Add agents subnet
Added a new subnet to network.bicep to correspond to the build agents subnet represented in the reference architecture article. The subnet will be initially empty but serves as a placeholder to quickly implement a build agent solution (for example: deploying an ACI resource) that would be able to publish the solution files directly to storage. Avoiding this way the need to manually update the storage firewall settings.